requirements.txt
Pillow==9.3.0 
beaker-pyteal==0.4.1
cffi==1.15.1
docstring-parser==0.14.1
msgpack==1.0.4
py-algorand-sdk==1.20.2
pycparser==2.21
pycryptodomex==3.16.0
PyNaCl==1.5.0
git+https://github.com/algorand/pyteal@tuple-element-encode#egg=pyteal
qrcode==7.3.1
semantic-version==2.10.0
sympy==1.11.1
